update release scripts

+++ libgringo/src/term.cc
@@ -561,7 +561,7 @@ bool Term::bind(VarSet &bound) {
 
 UTerm Term::insert(ArithmeticsMap &arith, AuxGen &auxGen, UTerm &&term, bool eq) {
     unsigned level = term->getLevel();
-    if (arith[level].find(term) == arith[level].end()) { level = arith.size() - 1; }
+    if (arith[level].find(term) == arith[level].end()) { level = numeric_cast<unsigned>(arith.size() - 1); }
     auto ret = arith[level].emplace(std::move(term), nullptr);
     if (ret.second) { ret.first->second = auxGen.uniqueVar(ret.first->first->loc(), level, "#Arith"); }
     if (eq) {
